+++ web/frontend_net_port/package.json
@@ -15,6 +15,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.8",
     "formik": "^2.4.5",
+    "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-cookie": "^7.1.4",
     "react-dom": "^18.2.0",
