+++ web/frontend_net_port/src/theme/TableTheme.js
@@ -1,0 +1,26 @@
+import TableCell, { tableCellClasses } from '@mui/material/TableCell';
+import TableRow from '@mui/material/TableRow';
+
+import { styled } from '@mui/material/styles'
+
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+    [`&.${tableCellClasses.head}`]: {
+        backgroundColor: '#bcc0ce',
+        color: theme.palette.common.black
+    },
+    [`&.${tableCellClasses.body}`]: {
+        fontWeight: theme.typography.fontWeightRegular
+    },
+}));
+
+const StyledTableRow = styled(TableRow)(({ theme }) => ({
+    '&:nth-of-type(even)': {
+        backgroundColor: theme.palette.action.hover
+    },
+    // hide last border
+    '&:last-child td, &:last-child th': {
+        border: 0
+    },
+}));
+
+export { StyledTableCell, StyledTableRow };
